telegram_bot/requirements.txt
python-dotenv==1.0.1
Requests==2.32.3
telepot==12.7
